+++ reversion/__init__.py
@@ -18,10 +18,6 @@ else:
         RevisionManagementError,
         RegistrationError,
     )
-    from reversion.signals import (  # noqa
-        pre_revision_commit,
-        post_revision_commit,
-    )
     from reversion.revisions import (  # noqa
         VersionAdapter,
         RevisionManager,