+++ CHANGELOG.rst
@@ -11,9 +11,9 @@ django-reversion changelog
 * Added ``reversion.views.RevisionMixin`` class-based view mixin (@etianen).
 * Revision blocks are now automatically wrapped in ``transaction.atomic()`` (@etianen).
 * Dramatically improved performance of version lookup for models with a non-integer primary key (@etianen, @mshannon1123).
-* Added ``ignore_duplicates`` and ``meta`` parameters to ``pre_revision_commit`` and ``post_revision_commit`` signals (@etianen).
 * Added ``get_for_model()`` function (@etianen).
 * Added support for using ``RevisionMiddleware`` with new-style Django 1.10 ``MIDDLEWARE`` (@etianen).
+* **Breaking:** Removed ``pre_revision_commit`` and ``post_revision_commit`` signals, use ``pre_save`` and ``post_save`` signals for ``Revision`` instead (@etianen).
 * **Breaking:** Removed ``get_for_date()`` function, use ``get_for_object().filter(revision__date_created__lte=date)`` (@etianen).
 * **Breaking:** Removed ``get_unique_for_object()`` function, use ``get_for_object().get_unique()`` instead (@etianen).
 * **Breaking:** Removed ``eager_signals`` argument to register(), it has been merged with the ``signals`` argument (@etianen).